Builds and sort of runs, but have verification issues with use of moving functions across modules
